docs: add JSdocs
