+++ packages/figma-utils/src/utils/fetch.ts
@@ -1,4 +1,4 @@
-import { FigmaVariablesApiResponse } from "./types.js";
+import { FigmaVariablesApiResponse } from "../types/figma.js";
 
 /**
  * Fetches the Figma Variables for the given file from the Figma API v1.
@@ -17,7 +17,7 @@ export const fetchFigmaVariables = async (
   });
 
   if (response.status !== 200) {
-    throw new Error("Figma API request failed");
+    throw new Error("Figma API request failed, status is not 200");
   }
 
   const data = (await response.json()) as FigmaVariablesApiResponse;
