+++ packages/figma-utils/src/utils/generate.ts
@@ -1,11 +1,34 @@
-import { ParsedVariable } from "./parse.js";
+import { ParsedVariable } from "../types/figma.js";
 
 type GenericGeneratorOptions = {
   data: ParsedVariable[];
+  /**
+   * Function which returns the full variable name depending on the used format.
+   *
+   * @example
+   * ```ts
+   * // for CSS:
+   * (name) => `--${name}`
+   * ```
+   */
   nameTransformer: (name: string) => string;
+  /**
+   * Function which returns the full variable value if its an alias / reference to another variable.
+   *
+   * @example
+   * ```ts
+   * // for CSS:
+   * (name) => `var(--${name})`
+   * ```
+   */
   aliasTransformer: (name: string) => string;
 };
 
+/**
+ * Generates the given variables into CSS variables.
+ *
+ * @returns File content of the .css file
+ */
 export const generateAsCSS = (data: ParsedVariable[]): string => {
   return genericGenerator({
     data,
@@ -14,6 +37,11 @@ export const generateAsCSS = (data: ParsedVariable[]): string => {
   });
 };
 
+/**
+ * Generates the given variables into Sass/SCSS variables.
+ *
+ * @returns File content of the .scss file
+ */
 export const generateAsSass = (data: ParsedVariable[]): string => {
   return genericGenerator({
     data,
@@ -22,7 +50,11 @@ export const generateAsSass = (data: ParsedVariable[]): string => {
   });
 };
 
-export const genericGenerator = (options: GenericGeneratorOptions) => {
+/**
+ * Generic base generator for CSS, SCSS etc. files.
+ * Will take care of defining selectors and formatting.
+ */
+const genericGenerator = (options: GenericGeneratorOptions) => {
   let content = "";
 
   options.data.forEach(({ modeName, variables }) => {
@@ -46,6 +78,10 @@ ${Object.entries(variables)
 
 /**
  * Checks whether the given variable value is an alias / variable reference to another variable.
+ * Alias values are enclosed by curly braces.
+ *
+ * @example "{your-variable-name}"
+ * @returns `isAlias` whether the variable is an alias and `variableName` the raw variable name without curly braces.
  */
 export const isAliasVariable = (variableValue: string) => {
   const isAlias = /{.*}/.exec(variableValue);
